requirements.txt
globox
numpy
pandas
Pillow
streamlit
supervision
ultralytics